\name{genes}
\docType{data}

\alias{genes}

\title{Genes Expressed in Foetal Liver Progenitor Cells and Adult Hepatocytes.}

\description{
These data are from a microarray study examining miRNA and mRNA expression in fetal liver progenitor cells and adult hepatocytes. The genes dataframe contains genes expressed in (but not necessarily regulated between) both cell types and contains the Entrez gene ID and gene symbol.
}

\usage{data(genes)}

\format{
A dataframe.
}

\references{
Kung, J. W. C. et al. MS in preparation.
}

\keyword{datasets}
