\name{mirs}
\docType{data}

\alias{mirs}

\title{miRNA Regulated Between Foetal Liver Progenitor Cells and Adult Hepatocytes.}

\description{
These data are from a microarray study examining miRNA and mRNA expression in foetal liver progenitor cells and adult hepatocytes. The mirs dataframe contains miRNA identifiers (e.g. hsa-miR-451) and fold change (log2 scale).
}
\usage{data(mirs)}

\format{
A dataframe.
}

\references{
Kung, J. W. C. et al. MS in preparation.
}

\keyword{datasets}
